Split plugin API to make it thread safe. #49 (#65)
Fixes https://github.com/rust-dsp/rust-vst/issues/49

This is a breaking change. If your plugin uses parameters or an editor,
you will need to update your code.

A step-by-step guide on how to port a plugin written for the old API to
the new API is available in the transfer_and_smooth example.